Fix Composites behavior (rendering and publishing)
From my understanding a Composite acts as a container for multiple Features (e.g. a time schedule ), which can be applied at once (e.g. a button click).

Issue 1: Modifications are not rendered (see e.g. issue #927).
Currently the modified state is stored (in this.state)  but not rendered. This leads to confusion and is not consistent with other Features where the modification is rendered immediately. For consistency the modification should be rendered immediately. This however leads to the question what should be rendered if there is a modification and the state of the device changes. I would still suggest that the modification is rendered.

Issue 2: Only modified features are published.
I think it would be beneficial to publish the entire Composite state (as a combination of modifications and device state). Not sure if this is correct and consistent accross the project.